doc: show a `git blame` equivalent in the git comparison
The repo's README says that blame is not implemented, so it makes
sense to not advertize an equivalent. But I've felt the need for it
the other day, and I figured out how to do it, so I thought: maybe I
should write that down, even if it's not convenient and of course
dependent on the git backend.